refactor: Reviews page is now search reviews
Have a page with just a long list of lecturers doesn't really provide
value to the user. The user will most likely search for the lecturer
or course they want to see reviews for. Having this one screen removes
the need for redux on lecturers since the search screen doesn't need to
be in sync.